move EditorLayout.tsx into layout/ directory
