+++ src/pageEditor/layout/EditorLayout.tsx
@@ -19,14 +19,14 @@ import React from "react";
 import ModListingPanel from "@/pageEditor/modListingPanel/ModListingPanel";
 import { useSelector } from "react-redux";
 import useFlags from "@/hooks/useFlags";
-import Modals from "./modals/Modals";
+import Modals from "../modals/Modals";
 import { selectIsInsertingNewStarterBrick } from "@/pageEditor/slices/editorSelectors";
 import EditorContent from "@/pageEditor/EditorContent";
-import styles from "./Editor.module.scss";
+import styles from "../Editor.module.scss";
 import RestrictedPane from "@/pageEditor/panes/RestrictedPane";
 import InsertPane from "@/pageEditor/panes/insert/InsertPane";
-import useCurrentInspectedUrl from "./hooks/useCurrentInspectedUrl";
-import NonScriptablePage from "./NonScriptablePage";
+import useCurrentInspectedUrl from "../hooks/useCurrentInspectedUrl";
+import NonScriptablePage from "../NonScriptablePage";
 import { isScriptableUrl } from "webext-content-scripts";
 import Loader from "@/components/Loader";
 import { selectIsStaleSession } from "@/store/sessionChanges/sessionChangesSelectors";
